+++ CHANGELOG.md
@@ -12,6 +12,8 @@ All notable changes to the ADTool will be documented in this file.
   - fixed issue when importing and no tree is open (issue #7)
   - fixed issue with comment not being imported from XML
   - improved error messages for command line options
+  - added option change number of ranked attacks
+  - added option to export or rank only chosen domains via command line
 
 ## [2.1.1] - 2016-04-28 ##
   - fixed not unfolding with space